twitter-producer/requirements.txt
TwitterAPI==2.4.5
python-dotenv==0.6.4
kafka-python==1.3.3
